Assignment 2
I'm struggling with this assignment, because I cannot figure out how to
save the CSS and Jquerry separately so that I may reference them in my
html later.  I am confident with  my html, but I still can't figure out
why it is not working.  I tried working with a peer, but I think she is
in the same boat as me.